Add (slightly fixed) scripts for getting input data
The FASTA input sequences for the GRC regions, as well as the MAF files
against which the alignments are compared and the BED files of genes
used to evaluate ortholog and paralog mappings, ought to be able to be
re-obtained with these scripts, provided you have hgsql set up and
pointed at a genome browser database.

Because I need to output a MAF from BioPython, I've set
trequirements.txt to install a BioPython incorporating the MAF support
PR, which they still haven't taken (because nobody has time to fix it up
I think).